requirements.txt
flask==2.2.3
flask-restful==0.3.9
flask-sqlalchemy==3.0.3
flask-jwt-extended==4.4.4
werkzeug==2.2.3
flask-cors==3.0.10
requests==2.28.2
python-dotenv==1.0.0

